Revert "StatisticsParams moved to proper folders"
This reverts commit 67d3514a0b460b0b40be4d4a1f9a843120b7fdb6.
