+++ apps/vaporgui/StatisticsParams.h
